deinline fmpq_poly IO
